+++ Sources/SwifterSwift/AppKit/NSViewExtensions.swift
@@ -109,6 +109,22 @@ public extension NSView {
         }
     }
 
+    /// SwifterSwift: Background color of the view; also inspectable from Storyboard.
+    @IBInspectable
+    var backgroundColor: NSColor? {
+        get {
+            if let colorRef = layer?.backgroundColor {
+                return NSColor(cgColor: colorRef)
+            } else {
+                return nil
+            }
+        }
+        set {
+            wantsLayer = true
+            layer?.backgroundColor = newValue?.cgColor
+        }
+    }
+
     /// SwifterSwift: Size of view.
     var size: CGSize {
         get {